Move io_uring support code and tests into the support/io_uring subfolders
