+++ src/program.c
@@ -21,8 +21,8 @@
 #include "log.h"
 #include "signal_handler_sigsegv.h"
 #include "memory_fences.h"
-#include "io_uring_support.h"
-#include "io_uring_capabilities.h"
+#include "support/io_uring/io_uring_support.h"
+#include "support/io_uring/io_uring_capabilities.h"
 #include "network/channel/network_channel.h"
 #include "network/channel/network_channel_iouring.h"
 #include "worker/worker.h"
