+++ src/worker/worker_iouring.c
@@ -10,8 +10,8 @@
 #include "xalloc.h"
 #include "thread.h"
 #include "memory_fences.h"
-#include "io_uring_support.h"
-#include "io_uring_capabilities.h"
+#include "support/io_uring/io_uring_support.h"
+#include "support/io_uring/io_uring_capabilities.h"
 #include "network/io/network_io_common.h"
 #include "network/channel/network_channel.h"
 #include "network/channel/network_channel_iouring.h"
