+++ tests/worker/test-worker-iouring.cpp
@@ -16,7 +16,7 @@
 #include "network/channel/network_channel.h"
 #include "network/channel/network_channel_iouring.h"
 #include "worker/worker.h"
-#include "io_uring_support.h"
+#include "support/io_uring/io_uring_support.h"
 
 #include "worker/worker_iouring.h"
